+++ cmd/repoctl/down.go
@@ -8,6 +8,7 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/cassava/repoctl/pacman/aur"
 	"github.com/cassava/repoctl/pacman/graph"
 	"github.com/cassava/repoctl/pacman/pkgutil"
 	"github.com/spf13/cobra"
@@ -15,6 +16,7 @@ import (
 
 var (
 	downDest     string
+	downDryRun   bool
 	downClobber  bool
 	downExtract  bool
 	downUpgrades bool
@@ -27,6 +29,7 @@ func init() {
 	MainCmd.AddCommand(downCmd)
 
 	downCmd.Flags().StringVarP(&downDest, "dest", "d", "", "output directory for tarballs")
+	downCmd.Flags().BoolVarP(&downDryRun, "dry-run", "n", false, "don't download any packages")
 	downCmd.Flags().BoolVarP(&downClobber, "clobber", "l", false, "delete conflicting files and folders")
 	downCmd.Flags().BoolVarP(&downExtract, "extract", "e", true, "extract the downloaded tarballs")
 	downCmd.Flags().BoolVarP(&downUpgrades, "upgrades", "u", false, "download tarballs for all upgrades")
@@ -40,13 +43,38 @@ var downCmd = &cobra.Command{
 	Aliases: []string{"download"},
 	Short:   "download and extract tarballs from AUR",
 	Long: `Download and extract tarballs from AUR for given packages.
-Alternatively, all packages, or those with updates can be downloaded.
-Options specified are additive, not exclusive.
 
-By default, tarballs are deleted after being extracted, and are placed
-in the current directory.
+  Alternatively, all packages, or those with updates can be downloaded.
+  Options specified are additive, not exclusive.
+
+  By default, tarballs are deleted after being extracted, and are placed
+  in the current directory.
+
+  Packages can also be downloaded recursively, and the list that these
+  dependencies should be built can be saved. For example, to download
+  all updates to the repository and build them in approximately the
+  correct order:
+
+    repoctl down -o build-order.txt -u
+    for pkg in $(cat build-order.txt); do
+        (
+            cd $pkg
+            makepkg -si
+            ok=$?
+            if $ok; then
+                repoctl add -m *.pkg.tar*
+                cd ..
+                rm -rf $pkg
+            fi
+        )
+    done
+
+  Caveat: Automatic dependency resolution does not currently handle version
+  resolution or library specifications, as noted in the Arch wiki at:
+  https://wiki.archlinux.org/index.php/PKGBUILD#Dependencies
 `,
-	Example: `  repoctl down -u`,
+	Example: `  repoctl down -u
+  repoctl down -o build-order.txt -u`,
 	RunE: func(cmd *cobra.Command, args []string) error {
 		// First, populate the initial list of packages to download.
 		var list []string
@@ -70,28 +98,51 @@ in the current directory.
 
 		// If no dependencies are wanted, then get to it right away:
 		if !downRecurse && downOrder == "" {
+			// There's not much point to a try run here, but we should respect
+			// the option nevertheless.
+			if downDryRun {
+				return nil
+			}
 			return Repo.Download(nil, downDest, downExtract, downClobber, list...)
 		}
 
-		// Otherwise, get the dependencies:
-		g, err := Repo.DependencyGraph(nil, list...)
+		// Otherwise, get the dependency list and download the packages:
+		aps, err := downDependencies(list)
 		if err != nil {
 			return err
 		}
-		_, aps, ups := graph.Dependencies(g)
-		if downOrder != "" {
-			f, err := os.Create(downOrder)
-			if err != nil {
-				return err
-			}
-			for _, p := range aps {
-				fmt.Fprintln(f, p.Name)
-			}
-			f.Close()
-		}
-		for _, u := range ups {
-			fmt.Fprintf(os.Stderr, "warning: unknown package %s\n", u)
+		// Don't download any packages if dry run is activated.
+		if downDryRun {
+			return nil
 		}
 		return Repo.DownloadPackages(nil, aps, downDest, downExtract, downClobber)
 	},
+}
+
+func downDependencies(packages []string) (aur.Packages, error) {
+	g, err := Repo.DependencyGraph(nil, packages...)
+	if err != nil {
+		return nil, err
+	}
+	_, aps, ups := graph.Dependencies(g)
+	if downOrder != "" {
+		f, err := os.Create(downOrder)
+		if err != nil {
+			return nil, err
+		}
+
+		for i := len(aps); i != 0; i-- {
+			fmt.Fprintln(f, aps[i-1].Name)
+		}
+		f.Close()
+	}
+	for _, u := range ups {
+		fmt.Fprintf(os.Stderr, "warning: unknown package %s\n", u)
+		iter := g.To(g.NodeWithName(u).ID())
+		for iter.Next() {
+			node := iter.Node().(*graph.Node)
+			fmt.Fprintf(os.Stderr, "         required by: %s\n", node.PkgName())
+		}
+	}
+	return aps, nil
 }